backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-environ
celery
redis
ulid-py
psycopg2-binary>=2.9
gunicorn
django-celery-beat
djangorestframework-simplejwt
whitenoise
aiofiles==24.1.0
aiogram==3.15.0
aiogram_dialog==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
cachetools==5.5.2
celery==5.6.0
certifi==2025.11.12
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.9
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.1
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
six==1.17.0
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ulid-py==1.1.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0